rename : myPage 폴더 이동
