+++ src/app/my/_components/MyPageContent/index.scss
